LPS-73562 Move SelfMonitorBundleListener out as a standalone class
